QoL upgrades
I added a little code to the notify icon so that when the form is minimized, it minimizes to the tray icon.  When you double click the icon, it will bring the main window back.
Again, it's not all pretty at this moment and mostly spaghetti code, but it's moving along little by little.